feat: working version
- remove functional test files
- rename routes -> controllers
- remove db session injection from the controller level
- packaged high level User logic into service
