--- ums/api/v2/controllers/group.py
@@ -1,6 +1,0 @@
-from pydantic import BaseModel
-
-
-class GroupPublic(BaseModel):
-    name: str
-    description: str | None
