+++ ums/api/v2/controllers/status.py
@@ -2,7 +2,7 @@ from typing import Annotated
 
 from fastapi import APIRouter, Depends
 
-from ums.api.v2.controllers.auth import get_current_active_user
+from ums.api.auth import get_current_active_user
 from ums.domain.entities import User
 
 router = APIRouter(tags=["status"])
