+++ ums/domain/group/schemas.py
@@ -1,0 +1,6 @@
+from ums.domain.data_access.interfaces import Schema
+
+
+class GroupPublic(Schema):
+    name: str
+    description: str | None
