--- ums/domain/user/_test_get_many.py
@@ -1,26 +1,0 @@
-import asyncio
-
-from loguru import logger
-
-from ums.core.filter_sort import SortParams
-from ums.db.async_session import get_async_session
-from ums.domain.user.reader import UserFilterParams, user_reader
-
-
-async def get_user_by_name():
-    result = await user_reader.get_many(
-        get_async_session,
-        UserFilterParams(
-            is_active=True,
-            is_verified=False,
-        ),
-        SortParams(
-            sort_by="full_name",
-            sort_order="asc",
-        ),
-    )
-    logger.info(result)
-
-
-if __name__ == "__main__":
-    asyncio.run(get_user_by_name())
